requirements.txt
python-telegram-bot==21.5
boto3
matplotlib
